+++ koala-domain/src/main/kotlin/comit/simsimpulyi/koala/domain/user/service/SecurityPort.kt
@@ -1,4 +1,4 @@
-package comit.simsimpulyi.koala.domain.user.spi
+package comit.simsimpulyi.koala.domain.user.service
 
 interface SecurityPort {
